requirements.txt
pytesseract==0.3.13
pdf2image==1.17.0
opencv-python==4.11.0.86
numpy==2.2.3
spacy==3.8.11
pandas==2.2.3
pydantic==2.12.5
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.2.1
pytest==8.3.5
scikit-learn==1.8.0
tqdm==4.67.3
python-dateutil==2.9.0.post0
google-generativeai==0.8.4
python-multipart==0.0.20
